Change how to determine host IP address
Previously we determined the host IP address by performing a UDP
connection and examining the default route used.  This does not work in
rootless because the cni plugin runs in a separate network namespace.
Now we look for the host IP address as an environment variable which is
reliable for rootful and rootless.

Signed-off-by: Brent Baude <bbaude@redhat.com>