Fix the cmdCheck command
cmdCheck was non functional as it always errors because of an invalid
gvproxy url. This command should contain some actual logic to check if
the ports are still working. To do so we get the port list from gvproxy
and compare it against our wanted ports. If the port is not found we
error (most likely something wrong in gvproxy).

Signed-off-by: Paul Holzinger <pholzing@redhat.com>